refactor(frontend): decompose agent studio model selection
